+++ java/compaction/compaction-core/src/test/java/sleeper/compaction/core/testutils/InMemoryCompactionJobStatusStoreTest.java
@@ -30,11 +30,6 @@ import java.time.Instant;
 import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static sleeper.compaction.core.job.CompactionJobEventTestData.committedEventBuilder;
-import static sleeper.compaction.core.job.CompactionJobEventTestData.defaultCompactionJobCreatedEventForTable;
-import static sleeper.compaction.core.job.CompactionJobEventTestData.failedEventBuilder;
-import static sleeper.compaction.core.job.CompactionJobEventTestData.finishedEventBuilder;
-import static sleeper.compaction.core.job.CompactionJobEventTestData.startedEventBuilder;
 import static sleeper.compaction.core.job.CompactionJobStatusTestData.compactionCommittedStatus;
 import static sleeper.compaction.core.job.CompactionJobStatusTestData.compactionFailedStatus;
 import static sleeper.compaction.core.job.CompactionJobStatusTestData.compactionFinishedStatus;
@@ -46,6 +41,11 @@ import static sleeper.core.record.process.status.ProcessStatusUpdateTestHelper.d
 import static sleeper.core.record.process.status.TestProcessStatusUpdateRecords.forJob;
 import static sleeper.core.record.process.status.TestProcessStatusUpdateRecords.forJobOnTask;
 import static sleeper.core.record.process.status.TestProcessStatusUpdateRecords.records;
+import static sleeper.core.tracker.compaction.job.CompactionJobEventTestData.committedEventBuilder;
+import static sleeper.core.tracker.compaction.job.CompactionJobEventTestData.defaultCompactionJobCreatedEventForTable;
+import static sleeper.core.tracker.compaction.job.CompactionJobEventTestData.failedEventBuilder;
+import static sleeper.core.tracker.compaction.job.CompactionJobEventTestData.finishedEventBuilder;
+import static sleeper.core.tracker.compaction.job.CompactionJobEventTestData.startedEventBuilder;
 
 class InMemoryCompactionJobStatusStoreTest {
