+++ java/core/src/test/java/sleeper/core/tracker/compaction/job/CompactionJobEventTestData.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package sleeper.compaction.core.job;
+package sleeper.core.tracker.compaction.job;
 
 import sleeper.core.record.process.ProcessRunTime;
 import sleeper.core.record.process.RecordsProcessedSummary;
